+++ autogpt_platform/frontend/src/components/DataTable.tsx
@@ -11,59 +11,12 @@ import {
 } from "./ui/table";
 import { Clipboard } from "lucide-react";
 import { useToast } from "./ui/use-toast";
+import { ContentRenderer } from "./ui/render";
 
 type DataTableProps = {
   title?: string;
   truncateLongData?: boolean;
   data: { [key: string]: Array<any> };
-};
-
-const VideoRenderer: React.FC<{ videoUrl: string }> = ({ videoUrl }) => {
-  const getYouTubeVideoId = (url: string) => {
-    const regExp =
-      /^.*((youtu.be\/)|(v\/)|(\/u\/\w\/)|(embed\/)|(watch\?))\??v?=?([^#&?]*).*/;
-    const match = url.match(regExp);
-    return match && match[7].length === 11 ? match[7] : null;
-  };
-
-  const videoId = getYouTubeVideoId(videoUrl);
-
-  return (
-    <div className="w-full p-2">
-      {videoId ? (
-        <iframe
-          width="100%"
-          height="315"
-          src={`https://www.youtube.com/embed/${videoId}`}
-          frameBorder="0"
-          allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-          allowFullScreen
-        ></iframe>
-      ) : (
-        <video controls width="100%" height="315">
-          <source src={videoUrl} type="video/mp4" />
-          Your browser does not support the video tag.
-        </video>
-      )}
-    </div>
-  );
-};
-
-const ImageRenderer: React.FC<{ imageUrl: string }> = ({ imageUrl }) => (
-  <div className="w-full p-2">
-    <img src={imageUrl} alt="Image" className="h-auto max-w-full" />
-  </div>
-);
-
-const isValidVideoUrl = (url: string): boolean => {
-  const videoExtensions = /\.(mp4|webm|ogg)$/i;
-  const youtubeRegex = /^(https?:\/\/)?(www\.)?(youtube\.com|youtu\.?be)\/.+$/;
-  return videoExtensions.test(url) || youtubeRegex.test(url);
-};
-
-const isValidImageUrl = (url: string): boolean => {
-  const imageExtensions = /\.(jpeg|jpg|gif|png|svg|webp)$/i;
-  return imageExtensions.test(url);
 };
 
 export default function DataTable({
@@ -81,24 +34,6 @@ export default function DataTable({
         duration: 2000,
       });
     });
-  };
-
-  const renderCellContent = (value: any) => {
-    if (typeof value === "string") {
-      if (isValidVideoUrl(value)) {
-        return <VideoRenderer videoUrl={value} />;
-      } else if (isValidImageUrl(value)) {
-        return <ImageRenderer imageUrl={value} />;
-      }
-    }
-
-    const text =
-      typeof value === "object"
-        ? JSON.stringify(value, null, 2)
-        : String(value);
-    return truncateLongData && text.length > maxChars
-      ? text.slice(0, maxChars) + "..."
-      : text;
   };
 
   return (
@@ -141,7 +76,10 @@ export default function DataTable({
                   </Button>
                   {value.map((item, index) => (
                     <React.Fragment key={index}>
-                      {renderCellContent(item)}
+                      <ContentRenderer
+                        value={item}
+                        truncateLongData={truncateLongData}
+                      />
                       {index < value.length - 1 && ", "}
                     </React.Fragment>
                   ))}
